feat(config): add presets support
fixes #173
