+++ packages/arui-scripts/README.md
@@ -291,7 +291,7 @@ docker run -p 8080:8080 container-name:version ./start.sh
 
 На `8080` порту будет поднят nginx, который будет раздавать статику и проксировать все остальные запросы к `nodejs`.
 
-Вы также можете переопределить полностью процесс сборки docker-образа, создав в корневой директории проекта `Dockerfile` содержащий необходимый набор инструкций. Пример [Dockerfile](src/templates/dockerfile.template.js).
+Вы также можете переопределить полностью процесс сборки docker-образа, создав в корневой директории проекта `Dockerfile` содержащий необходимый набор инструкций. Пример [Dockerfile](src/templates/dockerfile.template.ts).
 
 archive
 ---
@@ -459,8 +459,57 @@ module.exports = {
 
 Для некоторых конфигураций определены несколько ключей, они будут применяться в том порядке, в котором они приведены в этом файле.
 
----
-## Разработка
-Проект разбит на два пакета: `arui-scripts`, собственно содержащий все конфигурации и `arui-scripts-test`, используемый
-для тестирования. Для работы с пакетами используется [yarn workspaces](https://classic.yarnpkg.com/en/docs/workspaces/).
-Для
+Пресеты
+===
+
+В случае, если вы хотите использовать определенный набор конфигураций и оверрайдов сразу в нескольких проектах - вам может
+помочь механизм пресетов. Он позволяет выносить конфигурацию и оверрайды в отдельный пакет.
+Для того чтобы использовать персеты на проекте вы должны указать в package.json:
+
+```json
+{
+    "aruiScripts": {
+        "presets": "my-company-presets"
+    }
+}
+```
+
+Как пресет должен быть указан путь до папки с общими настройками (для поиска пути будет использоваться `require.resolve`
+от папки, содержащей package.json. Так что это может быть как папка в проекте, так и пакет из node_modules).
+
+Сам пакет с пресетами может содержать два файла:
+- `arui-scirpts.config.js`
+- `arui-scripts.overrides.js`
+
+### arui-scripts.config.js
+С помощью этого файла можно задать любые ключи [конфигурации](#настройки).
+```js
+module.exports = {
+    baseDockerImage: 'my-company-artifactory.com/arui-scripts-base:11.2'
+};
+```
+
+На проекте конфиурация будет загружаться в следующем порядке:
+1. базовые настройки из arui-scripts
+2. настройки из presets
+3. настройки из package.json проекта
+4. настройки, переданные через env переменную.
+
+**Важно!**
+
+Если вы будете задавать относительные пути через общие конфигурации (например `serverEntry`, `additionalBuildPath` и другие)
+они будут вычисляться относительно корня проекта, а не вашей конфигурации.
+Вы можете использовать абсолютные пути при необходимости задать путь до файла внутри пакета с пресетами.
+
+### arui-scripts.overrides.js
+С помощью этого файла можно задать базовые оверрайды проекта, аналогично [заданию оверрайдов на проекте](#тонкая-настройка).
+```js
+module.exports = {
+    babelClient(config) {
+        config.plugins.push('my-awesome-babel-plugin');
+        return config;
+    }
+};
+```
+
+На проекте оверрайды из пресетов будут выполняться в первую очередь, после них будут выполняться оверрайды из проекта.
