+++ packages/arui-scripts/src/configs/app-configs/available-package-settings.ts
@@ -1,0 +1,31 @@
+import { AppConfigs } from './types';
+
+/**
+ * Эти ключи из конфига будут обновляться из package.json при их наличии
+ */
+export const availablePackageSettings: Array<keyof AppConfigs> = [
+    'dockerRegistry',
+    'baseDockerImage',
+    'serverEntry',
+
+    'buildPath',
+    'assetsPath',
+    'additionalBuildPath',
+    'nginxRootPath',
+    'runFromNonRootUser',
+    'archiveName',
+
+    'serverOutput',
+    'clientPolyfillsEntry',
+    'clientEntry',
+    'keepPropTypes',
+
+    'useTscLoader',
+    'useServerHMR',
+    'clientServerPort',
+    'serverPort',
+
+    'debug',
+    'componentsTheme',
+    'keepCssVars',
+];
