+++ packages/arui-scripts/src/configs/app-configs/update-with-env.ts
@@ -1,0 +1,20 @@
+import merge from 'lodash.merge';
+import validateSettingsKeys from './validate-settings-keys';
+import { availablePackageSettings } from './available-package-settings';
+import { AppConfigs } from './types';
+
+export function updateWithEnv(config: AppConfigs) {
+    if (!process.env.ARUI_SCRIPTS_CONFIG) {
+        return config;
+    }
+
+    try {
+        console.warn('Используйте ARUI_SCRIPTS_CONFIG только для отладки');
+        const envSettings = JSON.parse(process.env.ARUI_SCRIPTS_CONFIG);
+        validateSettingsKeys(availablePackageSettings, envSettings);
+        return merge(config, envSettings);
+    } catch (e) {
+        console.error(e);
+        throw Error('Not valid JSON passed. Correct it. For example: ARUI_SCRIPTS_CONFIG="{\"serverPort\":3333}"');
+    }
+}
