+++ packages/arui-scripts/src/configs/app-configs/update-with-package.ts
@@ -1,0 +1,11 @@
+import merge from 'lodash.merge';
+import { AppConfigs } from './types';
+import validateSettingsKeys from './validate-settings-keys';
+import { availablePackageSettings } from './available-package-settings';
+
+export function updateWithPackage(config: AppConfigs) {
+    const packageSettings = config.appPackage.aruiScripts || {};
+
+    validateSettingsKeys(availablePackageSettings, packageSettings);
+    return merge(config, packageSettings);
+}
