+++ packages/arui-scripts/src/configs/app-configs/validate-settings-keys.ts
