+++ packages/arui-scripts/src/configs/util/__tests__/apply-overrides.tests.ts
@@ -1,6 +1,6 @@
 jest.mock('../../app-configs', () => ({
     hasOverrides: true,
-    overridesPath: 'overrides'
+    overridesPath: ['overrides']
 }));
 
 beforeEach(() => {
