+++ packages/arui-scripts/src/configs/util/apply-overrides.ts
@@ -1,9 +1,17 @@
 import appConfigs from '../app-configs';
 
-let overrides: Record<string, (config: any) => any> = {};
+type OverrideFile = Record<string, (config: any) => any>;
+
+let overrides: Array<OverrideFile> = [];
 
 if (appConfigs.hasOverrides) {
-    overrides = require(appConfigs.overridesPath);
+    overrides = appConfigs.overridesPath.map(path => {
+        try {
+            return require(path)
+        } catch (e) {
+            return {};
+        }
+    });
 }
 
 /**
@@ -17,12 +25,14 @@ function applyOverrides<T = any>(overridesKey: string | string[], config: T): T 
         overridesKey = [overridesKey];
     }
     overridesKey.forEach(key => {
-        if (overrides.hasOwnProperty(key)) {
-            if (typeof overrides[key] !== 'function') {
-                console.error(`Override ${key} must be a function`);
+        overrides.forEach((override) =>{
+            if (override.hasOwnProperty(key)) {
+                if (typeof override[key] !== 'function') {
+                    console.error(`Override ${key} must be a function`);
+                }
+                config = override[key](config);
             }
-            config = overrides[key](config);
-        }
+        });
     });
 
     return config;
