+++ packages/arui-scripts/src/configs/util/get-polyfills.ts
@@ -1,4 +1,4 @@
-import { AppConfigs } from '../config-type';
+import { AppConfigs } from '../app-configs/types';
 
 // require.resolve вынесен как параметр функции для того, чтоб это можно было протестировать.
 // на данный момент jest не дает возможности мокать require.resolve https://github.com/facebook/jest/issues/9543
