+++ packages/arui-scripts/src/configs/util/try-resolve.ts
@@ -1,0 +1,7 @@
+export function tryResolve(...args: Parameters<typeof require.resolve>) {
+    try {
+        return require.resolve(...args);
+    } catch (e) {
+        return undefined;
+    }
+}
